Added explicit click requirement to view active sessions, needed to prevent long load on loaded systems. Also added ACTIVE_SESSION_MINUTES parameter in Knar, to set the amount of time a session is considered active. Defaults to SessionExpire.
The expire program will need to be run pretty often, should add an
"expire sessions" action to this page or somewhere else. Actually,
we need a real cron interface, I am guessing, to run tasks like
that. Obviously we don't want to run crontab itself, but the daemon
ought to be able to fork long-running tasks.