Mike's DBConfig mechanism.
This is Mike's code that he was unable to merge on his own, so I did it.
I have not tested it (and I don't care if I am commiting untested code to CVS .... I didn't write it).

If you have problems with it talk to Mike, or check out the prvious versions of these files.

Talk to you all after the Thanksgiving.