Extensive changes to support new matrix and modular options.
  * Added [item-sku] tag to get base SKU for a variant

  * added fall-through for sku to [item-field]

  * rerouted [item-data ...] to always point to sku, not the code
    for the variant. (This will mean inventory has to be checked
    via a [data ...] tag when variants are in use.)

  * added [PREFIX-options] to get and parse options of all types.

Added profile setting to easily switch between pricing and shipping
types. Set in the $Session->{Autoload} array as a "tag-profile" scalar
of pattern \w+-\w+. When that is seen, the profile is read and loaded
if it exists. Profile sets various catalog.cfg directives; intended
to set CommonAdjust, PriceField, etc. with ease.