Add [mail ...] tag as standard. Does good header handling, supports raw mode in case of future MIME support.
Added a routine to support the auto-generated email copies of
form submissions from ::update_data.

More fixes to tag_accessories -- that thing has to be completely
rewritten....

Fix text2html to work with \r only.

Add pedantic=1 option to tag_tree to generate errors if an endless
tree is detected; otherwise it will tolerate (possibly logging) the
error and ignore the brance. Fixed endless tree handling so that it will not
error out unless the branch can actually be followed.