Add new tag "assign". It allows you to assign to four things,
	subtotal    preempts the cart subtotal derived from prices. NOT
			    ROUNDED.
	shipping    if any  non-zero mv_shipmode is present at all,
			    sets the total value of shipping. Rounded
			    to fractional digits.
	handling    if any non-zero mv_handling is present at all, sets the
	            total value of shipping. Rounded to fractional
			    digits.
	salestax    preempts the salestax normally derived from the
			    salestax. NOT ROUNDED.

You cannot assign to total_cost -- it will always be the sum of the
four above.

If there is no assignment to one of the four, it will use the
normal method. To make salestax zero, you must use [assign salestax=0].

Called with [assign salestax=N.NN], etc. If the value has a length
of zero, will delete the assigned value from the assignment hash
and the normal method of calculation will kick back in.

If you call [assign clear=1] it will clear all, and the normal method
of calculation will kick back in.

If a non-real-number value is passed, it logs an error and clears
the assignment.

It is persistent in the user's session and effects only that user,
and should be used only when you know exactly what you are doing. Ha.