Major overhaul of the way UI images are handled.
Instead of the old @_UI_IMG_@admin/ or @_UI_IMG_@navigation/ now all
images are in a single directory, and ready to be localized (thanks for the
hint, Stefan :). Images will be in a directory like this:

share/interchange/en_US

and so on depending on locale. And UI_STD_HEAD now modifies ImageDir so
that the image rewriting in Interpolate.pm takes care of the location
for us.

However, this can't handle JavaScript rollovers, etc., nor UI image
access from the regular catalog, and $CGI::secure isn't available inside
Safe, so I've added the [image] usertag, for getting the image directory
(and eventually doing other handy things).

This also meant I had to move std.js inline into UI_STD_HEAD, under
Interchange's control, to deal with image locations.

This nice thing is that in the UI, images are just called with <img
src="imgname.gif"> instead of the complicated stuff from before.

It remains to be seen how many other things I messed up along the way. :)