Add support for here documents in locale configuration with this setting:
MV_LOCALE_CHOMP_VALUES

The only problem with here documents previously was that (inherent in Perl)
they had to end in a newline, which we don't want. But the good thing about
here documents is they're the only way to put double-byte characters (or
anything moderately complex) into files without backslashes that make it
impossible to see e.g. Japanese characters in a text editor.

Example coming up.