merged changes 2.7.2.10 vs 2.7.2.11 +++ 2.7.2.11 +++ RATIONALE AND WARNING
Patch for the problem detailed below.

Mike said to the correctness of this patch a while ago:

I don't know. If the reset_calc doesn't happen before that, then
yes. Make sure that if you put it in, we don't do it again later.

Due to the slow release cycle of Interchange STABLE I put it in
now, so we try if it breaks sth. On my applications it worked
without any problems for quite a while.

PROBLEM DESCRIPTION

First case:

Autoload <<EOR
[perl]
...
[/perl]
EOR

Catalog and global usertags are working fine.

Second case:

Sub <<EOS
sub initialize {
....
}

Autoload initialize

Global usertags are working fine, but catalog ones not:

Tag 'cartadd' not defined.