Patch from Mark Johnson, whose explanation I quote:
If you send a garbage ORIGID, as IC does by default, it will attempt to
use that as the key for the transaction, without falling back on the data
you sent with it. Thus, it must explicitly not be used unless you have a
valid one which you want to use. This was why no credits were succeeding;
even though we were sending explicit new data, it ignored it and cried
with the ever useful error "Not signed up for this tender type".