Add new database configuration option PREFER_NULL. It accepts a list of field names which should be set to NULL instead of an empty string whenever possible:
Database  people  PREFER_NULL  age height

Probably most useful with NUMERIC fields that are nullable, to prevent
import errors when trying to store '' into a numeric field. Also for
character fields that you'd rather have contain a NULL than '' to comply
with a CHECK constraint or a foreign key constraint.

I tried to keep this as efficient as possible since these are some
frequently-called routines. If anyone can think of a faster way to handle
the quote sub, let me know.