Add new UserTag option "attrDefault", which allows default attributes to be set globally or catalog-wide for a given usertag. Two examples:
UserTag area attrDefault href index
UserTag calc attrDefault filter entities

Any user-specified attributes will take precedence, even if 0 or blank.

This is designed to allow changing the default behavior of a tag without
changing its code, especially for built-in tags one would rather not
customize. Note that when using this directive on global usertags,
it must be done in interchange.cfg, and for catalog usertags in catalog.cfg.