Fix rare but nasty bug that causes chained ITL conditional tests to fail in no-op mode:
	[if scratch something][or value something]...[/if]
	[if scratch something][and value something]...[/if]

Tests included.

The problem happens when the first [if] contains a string that doesn't
convert nicely to a number, and gets passed to an XOR test directly instead
of being converted to a 1 or 0 first. That causes a string XOR to be done,
which gives the wrong answer.

Since this is a sensitive area of code, I'd appreciate others checking it
out.