Flesh out alternate values spaces functionality Mike added on 2002-11-18.
In addition to his mv_values_space CGI setting, there is now a
[values-space] tag for setting the values space for the lifetime of the
current page (its various options are documented in the tag itself), and
[value] and [value-extended] both accept the values_space option to pull
a value out of an alternate space without changing the current one. The
name of the current values space is now stored in $Vend::ValuesSpace,
which is used when optionally copying values from the old to the new
values space in [values-space].

Tests included.

Also added enable_itl option to [value] to make it feature compatible with
[value-extended].