Trimmed the license to remove unneccesary text surrounding the code.
This is first change in practice according to interchange-core thread:
http://www.icdevgroup.org/mailman/private/interchange-core/2004-September/004216.html

I think the GNU GPL license does not force the whole license chapter to
be included in each file. This is what the license says:

[quote]
If you develop a new program, and you want it to be of the greatest possible
use to the public, the best way to achieve this is to make it free software
which everyone can redistribute and change under these terms.
To do so, attach the following notices to the program. It is safest to attach
them to the start of each source file to most effectively convey the exclusion
of warranty; and each file should have at least the "copyright" line and a
pointer to where the full notice is found.
[/quote]