Add a new database configuration parameter, POSTEXPORT, which allows calling one or more subs or globalsubs after each time that database is exported.
Sub is called with database name, exported filename (absolute path), and
an options hash consisting of the options given to the export call and
values for 'delim' and 'record_delim' for structuring the file pointed
to by filename.

For example, it could be used to send an email notification, make a log
file entry, or do postprocessing on the exported text file.

Written and contributed by Ethan Rowe <ethan@endpoint.com> (with a little
tweaking by me).