* Fix [query] causes [more] to generate non-unique mv_cache_key.
The [more] tag generates a unique mv_cache_key for [list], [search-region],
etc., but when used with [query], it never does.  This is because it is
using an incorrect "message" for generating the key: it uses the first 100
characters that come after the closing bracket of "[query]", which are never
unique.

Here is a patch that causes it to use the "sql" (e.g. "select * from ...") as
the unique message.

Here is an example of how to reproduce the problem:

[query ml="5" more=1 type=list
       sql="select * from products where prod_group = 'Hand Tools'"
]
    [on-match] [more-list] [more] [/more-list] <br> [/on-match]
    [list] [sql-param sku]<br> [/list]
[/query]

This will be used as the message:

    "[on-match] [more-list] [more] [/more-list] <br> [/on-match]
    [list] [sql-param sku]<br> [/li"

Whereas for a normal search (e.g. [loop]), it would use something like this:

   scan/fi=products/sf=prod_group/op=eq/se=Hand%20Tools/st=db/ml=5

Apparently, $page is whatever comes after [query], and that is never unique for
each query.  I have confirmed this bug in 4.9.9, 5.2.0, and CVS 2004-10-26.