Updates to the Linkpoint payment module from Ron Phipps <rphipps@reliant-solutions.com>. His comments:
I have fixed up Linkpoint.pm to be compatible with the new Linkpoint
Perl API (v6.0).  This is not backwards compatible with Linkpoint Perl
API (v5.4.1).  Linkpoint users are being forced to upgrade by Feb 28
2005.

There are a few outstanding issues in this module.  First the latest API
release from Linkpoint has an issue with the "scompany" and "saddress1"
fields.  Neither one of these will show up in Linkpoint Central.  The
suggested fix is to send Shipping Address Line 1 and Shipping Address
Line 2 in the "saddress2" field.  Once the API is fixed we will need to
remove this kludge from the payment module, there currently is no public
time frame for when this will happen. The Shipping Company is currently
being sent in "scompany" and once they fix the API it will show up
properly in Linkpoint Central.

The other issue is that the new Perl API allows you to call two
different methods to contact linkpoint.  You can either call it this
way:

my %result = $lperl->curl_process(\%query);

This uses curl to process the request.  Or you can call it this way:

my %result = $lperl->process(\%query);

This uses shared ssl libraries (liblperl.so and liblpssl.so).

The curl method just worked out of the box for me so that's the method
that is in use in Linkpoint.pm.