Re-order writing of foreign and password fields.
When you create a new account, it calls $udb->set_field first for
password, then it calls it again for the indirect login "foreign"
column if the indirect login was specified. The problem with this:
it creates a race condition in which a record momentarily exists with
a blank value in that foreign column, which is treated by Interchange
as if it has a unique constraint upon it (which it well may if it's
a foreign key constraint).

Therefore, this reverses the order in which these things happen, so
the password is populated second. If indirect_login isn't specified,
then there's no practical difference.

Really, there's a larger problem in how this is set up, because this
approach doesn't permit you to have NOT NULL constraints on columns in
the user table. This should probably be re-written so it can populate
the entire record at once instead of in several steps leaving some fields
empty at first.

Patch from Ethan Rowe <ethan@endpoint.com>.