Patch from Ethan Rowe <ethan@endpoint.com>:
Test 60 was failing due to an oversite on my part in
Vend::Interpolate::subtotal(). The pre-discount-space logic had a check:
my $discounts = defined $Vend::Session->{discount};

Later on, while looping through the products, it checks if $discounts is
true, and if so, adds the item's discounted subtotal to the running subtotal;
otherwise, it adds the item's full-price subtotal.

The discount spaces expect the $::Discounts hash to always exist. Therefore,
I changed the check to be based on:
my $discounts = (defined $::Discounts and %$::Discounts);

The discount_price sub forces the discount hash into exist when it encounters
an item with an mv_price attribute. However, the logic is now expecting the
discount hash to actually contain members, because it will always exist.
Therefore, the following patch to Interpolate.pm struck me as the right fix.
Let me know if you think there's a better way to do it. In any case, it
fixes the problem specific to test #60.