Add new pragma: no_default_reparse.
It makes the default Reparse setting false, which can be (as always)
overridden by the tag's Reparse setting, and finally by the reparse=X
tag attribute.

I almost never use reparse in any form, and thus prefer, for speed and
expectation purposes, to have it not be the default. If it's still wanted,
it can be turned on for that particular use only.