Use more generic SQL_NUMERIC instead of SQL_INTEGER to bind numeric parameters.
This fixes a problem on PostgreSQL >= 7.4 and recent versions of DBD::Pg
that now use server-side prepared statements and binding instead of
binding in the DBD itself. The PostgreSQL server was rejecting numbers
with decimals when binding to SQL_INTEGER, because, well, they're not
integers.

Tested on MySQL 3.23, PostgreSQL 8.0, and Oracle 8i.