Add dbi_quote filter, which uses DBI's (or the DBD's, if it overrides) quote method to make a string safe to use in SQL. It takes into account any database-specific quoting needs (such as \ for PostgreSQL or MySQL, truncating at the first \0 for PostgreSQL, turning a newline into a literal two-character '\n' for MySQL, etc.).
The database handle used is that of the first ProductFiles database,
unless another is specified with e.g.:

    [filter dbi_quote.inventory]some string or other[/filter]

Which is likely to result in:

    'some string or other'

Note the containing quotes, which is different than Interchange's native
"sql" filter. See also the DBI man page details about the quote method.

Since this uses database handles, Safe must be considered if it's being
used via the $Tag object in a Perl block.