Add new CartTrigger functionality to the core, by Ethan Rowe.
The CartTrigger functionality allows for specification of any number
of subroutines (global or catalog subs, specified by name) to execute
whenever the contents of a shopping cart are changed via the standard
means available through CGI variable space (i.e. when changes are
invoked via the 'process' system actionmap through mv_order_item and
mv_order_quantity field submissions, or from a standard Interchange cart
page). The subroutines will be executed per-change, such that any page
process resulting in multiple alterations to the cart will potentially
call these functions multiple times.

Directives are used to turn this behavior on/off and to control some
details of its behavior:

CartTrigger: an array-type directive; list the different subroutine
names to execute at cart-modification time here.

CartTriggerQuantity: a boolean-type (i.e. yes/no) directive, defaulting
to 'no'/'false'; when set to yes/true, changes to item quantity on an
existing cart member will cause the cart trigger subs to fire. A value
of 'no' (the default) means that quantity changes on existing cart lines
do not call the trigger (though a quantity change to zero will result in
a delete, which does fire the cart trigger).

Each subroutine specified in CartTrigger will be passed the following
arguments whenever they are called:
1. Reference to the cart
2. Scalar representing the action that fired the trigger; value will be
one of: add, update, delete
3. Hashref pointing to the new row (except for the 'delete' action, in
which case this will be undefined)
4. Hashref representing the old row (except for the 'add' action); for
'update' actions, this will be a *copy* of the row prior to the changes.
The old row is no longer a member of the cart.
5. The cart's name

The return value from each subroutine call is pushed onto an array;
when the particular trigger firing is complete (i.e. all subroutines
specified in CartTrigger have been called), the full array of results is
returned. However, the initial version of this functionality does not
use these return values in any meaningful way.

It must be noted that the Interchange cart subsystem is based
on arrayrefs of hashrefs; there is no object encapsulation for
limiting/monitoring programmatic access to the contents of any cart.
Consequently, direct manipulation of the cart from within Perl will
*not* cause these triggers to fire. The triggers only fire when the
cart contents are modified through the standard Interchange CGI-based
variables/processing. Therefore, it is assumed for now that any
programmer sufficiently comfortable/confident to manipulate cart
contents directly can also be given the responsibility of deciding
whether or not it is appropriate to invoke any cart triggers in response
to such changes.