Addition of "postlogin_action" profile behavior/argument for UserDB by Ethan Rowe <ethan@endpoint.com>.
Can provide a whitespace-separated list of sub names to execute
following a successful login; this "postlogin_action" list
directive/argument can be specified at the profile level or as an
argument to the [userdb] tag at login time. The subs will be executed in
the order they are listed within the "postlogin_action" value. They are
not passed any arguments, and their return values are not used in any
meaningful way; any behaviors desired from the login must therefore be
entirely encapsulated within the sub(s) specified in the list.