A patch from Ethan Rowe <ethan@endpoint.com>:
Add a new "size_scratch_prefix" argument to the [image] tag. If
"getsize" is in use (which is the default), passing a value for
"size_scratch_prefix" to [image] will result in the height and width
of the image being placed into temporary Scratch-space variables with
names "<size_scratch_prefix>_height" and "<size_scratch_prefix>_width",
respectively. Naturally, if "getsize" is given a perl-false value,
the dimensions will not be found for the image and thus the
"size_scratch_prefix" value is ignored.

(Dumb) Example:
    [image src=foo.png size_scratch_prefix=foo]
    foo.png has height [scratch foo_height] and width [scratch foo_width]

This kind of thing can be useful for styling control and page layout, as
it allows the application programmer to work with the image dimensions
and adjust markup/styling appropriately in response to those dimensions.

The scratch values created are temporary (i.e. as if they were created
via [tmp ...]) and will not persist in the session past the end of the
current page process.