Fix a bug found by Brian Miller <brian@endpoint.com>:
When reading an OrderProfile from an external config file,
for instance, etc/profiles.login or similar if there is a comment line
immediately preceding the __NAME__ identifier then the first line of the
profile is commented out. Actually any line preceding the __NAME__ line
will silently be prepended to the first line of the profile, just so
happens a comment (in some of our code) would be the likely offender.

For example:

# following profile confirms user input
__NAME__ Login
  username=required Username had better be filled in.
  password=required Password is required.
__END__

Will result in a profile structure that looks like:

# following profile confirms user input   username=required ...

And the username check will never execute.