Add new date_change_null filter.
This is meant for use with databases such as PostgreSQL that do not accept
invalid dates (the '0000-00-00' used in the date_change filter), but
require a NULL to indicate the absence of a date.

To use, set metadata to use widget "date_blank", filter "date_change_null",
and optionally set the default value to "0000-00-00" if you want the
default to be no date, rather than today's date.

Provided by Mark Johnson <mark@endpoint.com> with help from Stefan Hornburg.