Fix [error] tag so it does not attempt to replace format codes other than %s in the text or std_label attributes.  The following test code demonstrates the problem:
[error name=test set="This is a test error"]
[error name=test std_label=|
    <a href="[area href=formhelp form='
        field=mytable::test
        returnto=this/page
    ']">mylabel</a>
| show_error=1]

In the above test the [area] tag translates this/page in the form attribute to
this%2fpage.  The error tag then sees the %2f as a float specifier and
translates it to this0.000000page, also since the error text itself is gobbled
up the error itself does not display since there is nothing for the later %s
code to match against.  This patch corrects that problem by using a regexp to
substitute the error text instead of an sprintf, and so only the first instance
of %s will be substituted.