Wait until the *next* line of the profile before undefining $And.
$And gets set if there is a &and or &or command with no further args (ie
chaining the line above and below together, not chaining two or more tests
together on the same line).  Currently $And gets unset before it has a chance
to affect the following line.  This results in the &and or &or being ignored
and the two lines treated as individual tests.  This patch fixes that by
skipping to the next iteration of the loop before $And is undefined.