* Working Email.pm. * Compatibility functions for send_mail(), mail() and tag email. * Should be a drop-in replacement.
There is another part of this commit, that I will add later today or
tomorrow. (Re-directions from traditional send_mail, mail and tag email
to code in this module).

I'd like to give you a chance for peer review or comments before we
actually switch to Email.pm.

Cya folks
-doc