Add new BounceReferrals catalog directive, a boolean option.
When disabled (the default), there is no behavioral change.

When enabled, GET requests to URLs with mv_pc or mv_source set to an
affiliate code are redirected to the same URL minus the affiliate code.

This keeps search engines that respect redirects from storing the affiliate
code-salted URL in their indexes, and helps them focus on the real resource
with a single URL instead of a multitude of salted links.

It has the side-effect of causing regular users following an affiliate
link to land on a URL with no affiliate code but instead with a session
ID in it, if they didn't have a session cookie already. There's no
easy way around this, and I consider it a separate issue from the
BounceReferrals concept that either happens one the first page with
BounceReferrals, or the second page without it. Either way, if session
IDs in URLs are a concern, it'll need a separate solution.

(Historically, many application servers always bounce the first request
to check for cookie support. Nowadays, many simply require cookies for
anything that needs a session. Interchange is different on both counts.)