The new Ship.pm seems to have a lot of problems.  Here is a patch which addresses the following issues:
* Fixed problem where only the last shipping policy will get stored if
the multi-line format is used in shipping.asc.

* Fixed problem where options are not converted and stored properly on
all shipping policies.

* Moved more code into the new process_new_beginning sub:
	* By passing a ref to the @shipping array I can now push new records
onto the array from inside the sub.
	* The above means that the old @list array need not be returned since
it's only purpose at that point was to push the record onto @shipping,
then immediately be overwritten by @new.
	* That also means that @new can be assigned to @list inside the sub
instead of outside it.
	* Since we no longer need to return @new, this frees up the return
value which I have used to return the $first value instead.
	* Because $first is returned we no longer need to flag that condition
by assigning "_newmode" as the shipping mode.

* I also removed an entire redundant elsif branch from read_shipping.
The regexp that the branch tested for would never cause the branch to be
executed because the branch above it tests for a regexp that matches the
same lines.

I don't think it's a big deal to change the process_new_beginning sub
the way I did because the sub itself is only about 5 weeks old, so the
likelyhood of someone having implemented code around it is extremely small.