New universal "hide" attribute.  You can now set the hide=1 attrbute for any tag to supress its output.
In order to preserve backwards compatibility with existing individual tags' hide
attributes tag output is not reparsed if this is set regardless of the setting
of interpolate or reparse attributes.  The reasoning behind this is existing
hide=1 attributes clear the output of the tag before the output is reparsed so
setting hide=1 on those tags which used to support it had the effect of not
reparsing output (because there was no output left to reparse).  This can be
tested with the following line of ITL:

[tmp bar]bar[/tmp][calcn hide=1]'foo[tmp bar][area foo][/tmp]'[/calcn]::[scratchd bar]

In my testing the above line of html outputs:
::bar

Without the hide=1 attribute to calcn it outputs something like (url changed to
protect the guilty):
foo::http://www.example.com/foo.html

So hide=1 is effectively preventing reparsing as well as hiding the output.