Add new synonym [if var ...] = [if variable ...] to match the [var] tag.
The invalid [if var ...] form was quietly accepted but always evaluated
to true, so this synonym is an important usability enhancement for any
expected [if var ...] that would evaluate to true during testing but
misleadingly appears to work.

Includes a unit test, and also tests for [if variable ...] and
[if global ...].

Thanks to Dan Collis-Puro <dan@endpoint.com> for the idea.