Allow default error labels to be set in the Locale.
The previous behavior was to use the variable name when no label was given. This change causes it to first check for a default label in the form of an error_label_${variable name} entry in the Locale configuration before falling back on just the variable name. As such, it is fully backwards compatible.

This is useful if you use Locale to override the error messages in core modules and also want to override the display of the label without setting up or modifying the profile everywhere the error may be displayed (with labels) on the site. For example:

Locale en_US <<EOF
{
# Override an error in Order.pm with one specific to this catalog.
"'%s' for item %s is not numeric/integer",
"We do not offer fractions of a product; however, the product is designed so that you may cut it to the exact dimensions needed. (The quantity you entered was '%s'.)",

# Also override the error label.
"error_label_mv_order_quantity",
"Quantity",

}
EOF