Make global ActionMap input consistent with catalog ActionMaps (RT #48).
This will break all global ActionMaps that expect the old calling
parameter of a URL with the action removed from the beginning.

I've manually reviewed all of the built-in global ActionMaps, and tested
some of them, and found no problems. But I won't be surprised if some
problems are found. More testing is warranted.